refactor(x): neutralize image share to x naming
